// get-aboard-next/src/client/models/PatchedNode.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type PatchedNode = {
    readonly node_id?: number;
    flow?: number;
    title?: string;
    x_pos?: number;
    y_pos?: number;
    description?: string;
    readonly created_at?: string;
    readonly updated_at?: string;
};